Refactor Model and Move tests to model/experimental/pointpillars
