+++ models/experimental/pointpillars/tests/pcc/test_ttnn_second.py
@@ -2,19 +2,21 @@
 
 # SPDX-License-Identifier: Apache-2.0
 
+import math
 import ttnn
 import torch
 import pytest
 from loguru import logger
 
 from tests.ttnn.utils_for_testing import assert_with_pcc
-from models.utility_functions import run_for_wormhole_b0
+from models.common.utility_functions import run_for_wormhole_b0
 from ttnn.model_preprocessing import preprocess_model_parameters
-from models.experimental.functional_pointpillars.reference.mvx_faster_rcnn import MVXFasterRCNN
-from models.experimental.functional_pointpillars.reference.second import SECOND
-from models.experimental.functional_pointpillars.tt.ttnn_second import TtSECOND
-import math
+from models.experimental.pointpillars.common import load_torch_model, POINTPILLARS_L1_SMALL_SIZE
+
+from models.experimental.pointpillars.tt.ttnn_second import TtSECOND
+
 from models.experimental.yolo_common.yolo_utils import determine_num_cores, get_core_grid_from_num_cores
+from models.experimental.pointpillars.tt.model_preprocessing import create_custom_preprocessor
 
 
 def interleaved_to_sharded(x):
@@ -32,66 +34,16 @@ def interleaved_to_sharded(x):
     return ttnn.reshard(x, shardspec) if x.is_sharded() else ttnn.interleaved_to_sharded(x, shardspec)
 
 
-def fold_batch_norm2d_into_conv2d(conv, bn):
-    if not bn.track_running_stats:
-        raise RuntimeError("BatchNorm2d must have track_running_stats=True to be folded into Conv2d")
-
-    weight = conv.weight
-    bias = conv.bias
-    running_mean = bn.running_mean
-    running_var = bn.running_var
-    eps = bn.eps
-    scale = bn.weight
-    shift = bn.bias
-    weight = weight * (scale / torch.sqrt(running_var + eps))[:, None, None, None]
-    if bias is not None:
-        bias = (bias - running_mean) * (scale / torch.sqrt(running_var + eps)) + shift
-    else:
-        bias = shift - running_mean * (scale / torch.sqrt(running_var + eps))
-
-    return weight, bias
-
-
-def create_custom_preprocessor(device):
-    def custom_preprocessor(model, name, ttnn_module_args):
-        parameters = {}
-        if isinstance(model, SECOND):
-            parameters["blocks"] = {}
-            for index, child in enumerate(model.blocks):
-                parameters["blocks"][index] = {}
-                for i in range(0, len(child), 3):
-                    conv_weight, conv_bias = fold_batch_norm2d_into_conv2d(child[i], child[i + 1])
-                    parameters["blocks"][index][i] = {}
-                    parameters["blocks"][index][i]["weight"] = ttnn.from_torch(conv_weight)
-                    parameters["blocks"][index][i]["bias"] = ttnn.from_torch(
-                        conv_bias.reshape(1, 1, 1, -1),
-                    )
-        return parameters
-
-    return custom_preprocessor
-
-
 @pytest.mark.parametrize(
     "use_pretrained_weight",
     [
         True,
     ],
 )
-@pytest.mark.parametrize("device_params", [{"l1_small_size": 32768}], indirect=True)
+@pytest.mark.parametrize("device_params", [{"l1_small_size": POINTPILLARS_L1_SMALL_SIZE}], indirect=True)
 @run_for_wormhole_b0()
-def test_ttnn_second(device, use_pretrained_weight, reset_seeds):
-    reference_model = MVXFasterRCNN(
-        pts_voxel_encoder=True,
-        pts_middle_encoder=True,
-        pts_backbone=True,
-        pts_neck=True,
-        pts_bbox_head=True,
-        train_cfg=None,
-    )
-    if use_pretrained_weight == True:
-        state_dict = torch.load("hv_pointpillars_fpn_sbn-all_4x8_2x_nus-3d_20210826_104936-fca299c1.pth")["state_dict"]
-        reference_model.load_state_dict(state_dict)
-    reference_model.eval()
+def test_ttnn_second(device, use_pretrained_weight, reset_seeds, model_location_generator):
+    reference_model = load_torch_model(model_location_generator)
     reference_model = reference_model.pts_backbone
 
     parameters = preprocess_model_parameters(
